move the test and add some determinism
